WinHttpSyncHttpClient: Fix proxy when using schemes other than HTTPS
There was no way to tell which scheme to use for the proxy settings, so
the user sets ClientConfiguration::scheme to HTTPS for making HTTPS
requests to S3, it also sets HTTPS for the proxy host -- which does not
make any sense. Given that, on some versions of Windows 10 (e.g. RS1),
by setting a proxy host with HTTPS when it *actually* works over HTTP,
the proxy fails.

This patch should fix issue #382

Signed-off-by: Paulo Alcantara <paulo.alc.cavalcanti@hp.com>