CryptoImpl: support LibreSSL again
While LibreSSL strives to be API compatible with OpenSSL,
it does so against the version when it was forked, 1.0.1g

LibreSSL defines its own LIBRESSL_VERSION_NUMBER but at
the same time it defines OPENSSL_VERSION_NUMBER to an
insane value, which causes issues now.

By detecting this case and setting a more reasonable
version number, we can compile against LibreSSL again.

Signed-off-by: André Draszik <git@andred.net>