Add ability to install a global SIGPIPE handler
On some operating systems, curl's installation raises SIGPIPE and
terminates the process. This behavior is not consistent and only happens
on _some_ linux distros (e.g. Ubuntu). We already set CURLOPT_NOSIGNAL,
however, there are some situations in which curl will still cause the OS
to raise a SIGPIPE regardless.