Add support for 2-digit years in RFC822 date format
Two digit years will be calculated as post Y2K.
So, for example, 80 will be 2080 not 1980. This should be OK for
the 99% of the cases since this class is used to parse dates sent by
AWS services. For the cases which this doesn't work. The 4-digit version
should be used instead.