Ensure TransferManager's life extends beyond TransferHandle callbacks
By controlling the way TransferManager is being created, in this case on
the heap using a shared_ptr, we can guarantee it lives at least until the
last callback has finished executing. Regardless if the application
destroys/releases its copy.With that in mind, this patch introduces a
Builder pattern to control the creation of TransferManager.

This fixes https://github.com/aws/aws-sdk-cpp/issues/574