Fix cmake build on Windows using generators other than Studio
Visual Studio adds the the configuration (Release, Debug) as a
subdirectory to where it outputs the binaries. This patch fixes the
assumption that on Windows cmake _always_ runs with Studio as its
generator. For example, Ninja builds are broken because the dlls are not
copies to the test binaries because there is no "Debug" directory.