Add lock to CurlHandleContainer for growing the current pool size aggressively.
Added lock to CurlHandleContainer for growing the current pool size in case multiple threads access the pool at the same time and we grow the size of pool with an aggressive way, as long as the pool size doesn't go over maximum.

The problem stems from the situation that 3 threads trying to acquire handles at the same time and the initial pool size is 1. assume we have T1, T2 and T3 stand for 3 threads.

When T1 gets the handle.
T2 blocks and tries to grow the pool to 2 (double)
T3 skips blocking and goes directly to Acquire() and get blocked.
After T2 growing the pool and get out, T3 will get a handle and T2 will block by Acquire().

If T1 and T3 will not release the handles and no other threads come to grow the pool. T2 would block forever.

So our way work around is, as long as a thread detect there is no resource available, it will try to grow it. even though other thread is growing it currently. The grow part will be protected by mutex though..