Make TransferManager own the threadpool & s3client exclusively
TransferManager must control the threadpool it is using in order to
deterministically join all worker threads in its destructor.
Additionally, it must control the lifetime of the s3 client it is
using under the hood for the same reason; to deterministically join
all the woker threads created by s3client's async operations.

This should fix issue #574, issue #647 and issue #651