TransferManager should NOT own the thread executor
In some cases, a worker thread is the only thing holding on to TM through
a shared_ptr. When TM is destroyed, it can't destroy the thread-executor
which the current worker thread is owned by. This would result in a crash.

Note: when the thread executor dies, it joins all its threads in the dtor.