Use dynamic_cast when deleting polymorphic types
Types that implement multiple interfaces, can be deleted via any of its
base interfaces that point to the object. However, only the first
interface points to the beginning of the object memory layout. The
second, third, etc. interfaces always point to an offset of the object.
When calling Free(..), we need to be pointing to the beginning of the
object memory. The C++ way to get to there is by doing a
dynamic_cast<void*>(). Clang/GCC do not require RTTI to be turned on
to do this. However, MSVC throws an exception for ANY use of
dynamic_cast.

See http://en.cppreference.com/w/cpp/language/dynamic_cast