Encode the path section of URLs according to RFC 3986
Some of the URI reserved characters are used in the path of newer services.
Our current URL encoder encodes those characters (for example '$') to their
percent-encoded form, which results in bad requests. For example,
API-Gateway does not recognizing the resource path and rejects the request.

However, we can't simply whitelist the reserved characters from getting
encoded, because as of this writing, SignatureV4 calculation relies on
encoding _every_ character that is not in the unreserved character set.

In summary, we have some sort of impedance mismatch. Services except URLs
to be encoded one way, while SignatureV4 expects the _same_ URL to be
encoded in a different way.