Unsigned payloads
S3 C2j model doesn't specify the authorizer to be v4-unsigned-body. This leads to all of its requests returning true when inheriting bool AmazonWebServiceRequest::SignBody().

Proposed solution is to change the ctor argument to an enum with values AlwaysSign, NeverSign and RequestDependent. With the default value of that enum argument set to RequestDependent.
In S3's ctor, the value should be set to NeverSign.
In AWSClient, the values AlwaysSign and NeverSign overrule the value returned by the request's SignBody().