Introduce aws_resource_prefix CLI flag to denote a prefix (#239)
Currently, when you need to run a subset of the tests, you pass the
-gtest_filter flag to filter them. However, this usually gets
interpreted as the resource prefix when creating AWS resources in
integration tests. This is because the tests assume the first argument
(if any) is the value of the prefix. By adding a key to that value, we
can easily distinguish between test filters and resource prefixes.