Adds support for tagging an EBS snapshot as part of the API call that creates the EBS snapshot
Mediapackage expands their service to FRA and CDG regions.

Added support for delete-application API and the ability for developers to set a homepage for their application. The homepage is a URL with more information about the application, for example the location of your GitHub repository for the application.

This release adds an API for adding a file directly to an AWS CodeCommit repository without requiring a Git client.