Remove AWSFunction.h
The purpose of this file was to provide custom allocator aware
std::function. However, allocator support in std::function has unclear
semantics, and there are technical issues with storing an allocator in a
type-erased context and later needing it for allocations in a copy-assignment
Moreover, allocator support for std::function was removed in C++17.

See https://github.com/aws/aws-sdk-cpp/issues/652