Bypass proxy settings when fetching credentials
To prevent an attacker from using a proxy to intercept credentials, we
must bypass any proxy settings specified by the user.
Note that by default the SDK does not automatically pick up proxy
settings from the machine.
The only exception to that rule is when the SDK is built with
IXMLHttpRequest2 as the HTTP client. IXMLHttpRequest2 always picks up
the default proxy settings on the machine, and there is no documented
way to bypass it. For that reason, the SDK uses WinHttp exclusively when
fetching credentials.