Replace JsonCpp with cJSON
Profiling workloads that return large JSON records pointed at JsonCpp
being a bottleneck. Replacing JsonCpp  with cJSON and introducing a
JsonView class (a lightweight read-only abstraction of the JSON DOM)
resulted in ~25% boost in parsing & model deserialization.

The improvement was most noticeable in the deserialization of
DynamoDB responses to Scan requests.