Make IsRetryableHttpResponseCode consistent across SDK
isRetryableHttpResponseCode in AWSClient.cpp and GetErrorForHttpResponseCode in CoreErrors.cpp are inconsistent with which Http Response Codes is retryable or not.
Resolving this by defining a single IsRetryableHttpResponseCode function.

Make retry strategy in AwsHttpResourceClient only retry two times and change the scale factor used in retry strategy in core tests to 1ms so as to speed up test runs.