ValueExists must return false if value is null (#374)
JsonView::ValueExists has the wrong semantics. It tests whether a key
exists rather than a value. That leads to a multiple-check usage
pattern; first check for existence, then check if the value is not null.

This patch fixes the semantics of ValueExists to return false if there's
a key with value |null|. It also introduces KeyExists which has the
previous semantics of ValueExists.