Add support for S3 select API
This release adds support via to the S3 select streaming API.
The API is implemented in C99 via libraries that are developed
by AWS as well.

The libraries are downloaded and built as part of the CMake configure
step. That can be disabled via the new switch -DBUILD_DEPS=OFF. The
switch is set to ON by default.