Fix Directory Creation Test with Long Dir Name
On non-Windows systems, uses pathconf(3) when available (using cmake's check_symbol_exsits function) to get the maximum
file name length in the current directory.

This prevents the test from failing when run on file systems
with different limitations (e.g. aufs)

Change the way to determine the existence of umask function the same way as pathconf.