Fix x-compiling for iOS & tvOS
Pass the right CMake args to the third party dependencies build project.
Determining whether we're cross-compiling was a bit tricky. Settled for
checking if CMAKE_OSX_ARCHITECTURES is specified.

Also, removed the CMAKE_INSTALL_LIBDIR and CMAKE_INSTALL_BINDIR
variables that were set in the third-party project. Those are dealt with
in the respective projects internally.