Add support for event-stream encoding & signing
We already had event-stream decoding business logic implemented for
service responses. This patch adds support for encoding a request as an
event-stream along with the Signature-v4 logic associated with it.
This patch will be later followed by changes to the code-generator to
create clients that take advantage of this implementation.