Add support to use the region specified in the aws config file
You can specify a profile name in ClientConfiguration. This will read
the config file created by the aws-cli ('aws configure' command).

This enables the dynamic behavior of which region to target without
recompiling the code.

The 'format' value in the config file is not used at the moment. Only
the 'region' is used.