Correctly encode ';' in URL
This issue came out from S3.
Using the key 'a;jsessionid=1' results in an error because 'jsessionid'
happens to be the name of a J2EE cookie. The encoding prior to this
patch is technically correct and works with S3 keys that have the same
format (i.e. 'a;b=1') but the specific string 'jsessionid' trips the
webserver. Most likely it gets stripped out and therefore the computed
signature does not match.