fix cmake with CMAKE_INSTALL_PREFIX need write permission
When build the sdk with `cmake -DCMAKE_INSTALL_PREFIX=/opt/aws-sdk-cpp`

Because we build and install the dependencies in this step, the command
would fail, because no writting permission to /opt folder.

When the sdk is build without `-DCMAKE_INSTALL_PREFIX`, it uses a local
folder to build the dependencies and copy the built dependencies to
their final location in the `sudo make install` stage.

This change will make both case to use local temp build folder.