Amazon VPC Flow Logs adds support for 1-minute aggregation intervals.
This feature ensures that an instance is patched up to the available patches on a particular date. It can be enabled by selecting the 'ApproveUntilDate' option as the auto-approval rule while creating the patch baseline. ApproveUntilDate - The cutoff date for auto approval of released patches. Any patches released on or before this date will be installed automatically.

This release enables AWS MSK customers to list Apache Kafka versions that are supported on AWS MSK clusters. Also includes changes to expose additional details of a cluster's state in DescribeCluster and ListClusters APIs.

Documentation updates for CloudFront

This release adds support for tagging Amazon WorkMail organizations.

Adding KVM as a support hypervisor

Updated ThrottlingException documentation to report that the error code is 400, and not 429, to reflect actual system behaviour.