This release allows you to create and manage tags for event buses.
Documentation updates for iotcolumbo

Announcing persistent file systems for Amazon FSx for Lustre that are ideal for longer-term storage and workloads, and a new generation of scratch file systems that offer higher burst throughput for spiky workloads.

AWS Snowball adds a field for entering your GSTIN when creating AWS Snowball jobs in the Asia Pacific (Mumbai) region.

This release allows you to create and manage tags for event buses.

Documentation updates for docdb