Support new pseudo region: "aws-global" in ClientConfiguration.
"aws-global" pseudo region tells SDK to use service global endpoint if there is any.
If the resolved region is "aws-global", the signing region will be set to "us-east-1". The
endpoint will be set to global endpoint if there is one, otherwise, the endpoint will be
calculated based on "us-east-1".

S3 is the only service supporting this feature at this moment. If an S3 resource is not located in the specified region, SDK will try to discover correct region and endpoint from the error and retry the request.

And followRedirects in ClientConfiguration has three options now: DEFAULT, ALWAYS and NEVER, in which DEFAULT means SDK will handle responses with status code 30x properly to support "aws-global" region.

Co-authored-by: Andrew Tang <tangandr@amazon.com>