Make generated CMake files valid when install paths are absolute
Problem
-------

If the install directories are given as absolute paths, the generated
CMake files break completely; in particular, on NixOS, they will try
to read headers from

/nix/store/somelonghash-aws-version-dev//nix/store/somelonghash-aws-sdk-cpp-version-dev/include

and find the library binaries in

/nix/store/somelonghash-aws-version-dev//nix/store/somelonghash-aws-sdk-cpp-version/lib

The part before the `//` is what is discovered as AWSSDK_ROOT_DIR.

Fix
---

If CMAKE_INSTALL_LIBDIR and friends are absolute, then they should not
have AWSSDK_ROOT_DIR prepended to them.

Unfortunately, this is tricky to do in a small patch, but on the
assumption that if one of these variables is absolute, then so are the
rest, I have picked AWSSDK_INSTALL_LIBDIR as the representative sample.

If this variable is unset, then AWSSDK_ROOT_DIR is forced to be "" so
that in can still be prepended to the install paths and get a valid
path as a result.