Add support for code review and recommendation feedback APIs.
This release is to support AWS Firewall Manager policy with Organizational Unit scope.

Amazon Redshift support for usage limits

This change adds a new field 'OptionalDeployment' to ServiceSoftwareOptions to indicate whether a service software update is optional or mandatory. If True, it indicates that the update is optional, and the service software is not automatically updated. If False, the service software is automatically updated after AutomatedUpdateDate.

Adding ServiceUnavailableException as one of the expected exceptions