Adding support for global write forwarding on secondary clusters in an Aurora global database.
Documentation updates for meteringmarketplace

You can now configure Amazon SES to send event notifications when the delivery of an email is delayed because of a temporary issue. For example, you can receive a notification if the recipient's inbox is full, or if there's a temporary problem with the receiving email server.

AWS Elemental MediaConvert SDK has added support for NexGuard FileMarker SDK, which allows NexGuard partners to watermark proprietary content in mezzanine and OTT streaming contexts.

Added a new ListHostedZonesByVPC API for customers to list all the private hosted zones that a specified VPC is associated with.

Added offset support for specifying the number of days to wait after the date and time specified by a CRON expression before running the maintenance window.

Documentation updates for support