Amazon EventBridge adds support for API Gateway as a target.
Documentation update to the introduction text to specify that this is the Amplify Console API.

Added the option to use IP addresses from an HTTP header that you specify, instead of using the web request origin. Available for IP set matching, geo matching, and rate-based rule count aggregation.

Amazon CloudWatch Events/EventBridge adds support for API Gateway as a target.

Documentation updates for cloudhsmv2

This release introduces the following set of actions for the EBS direct APIs: 1. StartSnapshot, which creates a new Amazon EBS snapshot. 2. PutSnapshotBlock, which writes a block of data to a snapshot. 3. CompleteSnapshot, which seals and completes a snapshot after blocks of data have been written to it.

AWS Comprehend now supports Real-time Analysis with Custom Entity Recognition.

Adds support for filters on the ListSecrets API to allow filtering results by name, tag key, tag value, or description.  Adds support for the BlockPublicPolicy option on the PutResourcePolicy API to block resource policies which grant a wide range of IAM principals access to secrets. Adds support for the ValidateResourcePolicy API to validate resource policies for syntax and prevent lockout error scenarios and wide access to secrets.

AppMesh now supports Ingress which allows resources outside a mesh to communicate to resources that are inside the mesh. See https://docs.aws.amazon.com/app-mesh/latest/userguide/virtual_gateways.html

This release adds the DeleteHumanTaskUi API to Amazon Augmented AI

Added support for registering an AVS device directly to a room using RegisterAVSDevice with a room ARN

This release adds support for SMS origination number as an attribute in the MessageAttributes parameter for the SNS Publish API.