Documentation updates for Amazon SNS
Support for HLS discontinuity tags in the child manifests. Support for incomplete segment behavior in the media output. Support for automatic input failover condition settings.

Adding DocDbSettings to support DocumentDB as a source.

Documentation updates for AWS ElastiCache

This release of the Amazon Macie API adds an eqExactMatch operator for filtering findings. With this operator you can increase the precision of your finding filters and suppression rules.

This release supports tagging for Amazon Braket quantum-task resources. It also supports tag-based access control for quantum-task APIs.

This feature increases the number of accounts that can be added to the Launch permissions within an Image Builder Distribution configuration.